Fix a 500 on the admin employees page, and the class behind it
`/admin/employees` answered **HTTP 500** for every listing containing an
employee with a contract duration. `EmployeeStore` read the column as
`(Integer) rs.getObject("contract_duration_months")`; it is `int(10) unsigned`,
whose range does not fit a signed `int`, so MariaDB Connector/J boxes it as a
`Long` and the cast threw `ClassCastException`.

Read with `getLong` plus `wasNull` instead, the way `LegacyJdbcValues` already
reads every numeric column on the legacy surface, and `Employee`'s field
becomes a `Long` to match the column's real range.

**Why 2,959 tests did not catch it.** Not one fixture in
`AdminEmployeesEndToEndTest` ever set that column, so the cast never ran. The
new regression test sets it to 12 and fails without the fix. 1,448 of the
development seed's 3,783 employees carry a value: the page failed on the first
real data it was shown.

**The class, not the instance.** `JdbcBoxedCastCoverageTest` forbids casting a
`getObject()` result to a box type anywhere under `com.workin`. The box type
the driver picks is a property of the column -- its width and its signedness --
and no reader of Java source can see it, so every such cast is a runtime
failure waiting for an unsigned column, passing review because the Java reads
correctly. Probing for null with an untyped `Object` and then reading through a
typed getter stays legal; that is what `AttendanceStore` already does.

The guard is proven against the shape it exists to catch, and it blanks
comments before scanning -- the first run flagged its own javadoc, which quotes
the defect. `check_client_endpoints_drift.py` had already learned that lesson
where a commented-out constant counted as a live endpoint.

Found by the deployment E2E sweep visiting every dashboard page against the
sanitised seed. No unit test, review or manual check had opened that page with
real data in it.

D-195.

+++ backend/src/main/java/com/workin/backend/platformadmin/hr/Employee.java
@@ -14,11 +14,17 @@ import java.math.BigDecimal;
  * <p>Neither the shift nor the salary is a column here. Both are the latest
  * row of a history table, picked per employee by a correlated subquery, so an
  * employee with no assignment or no contract shows nothing rather than a zero.
+ *
+ * <p>{@code contractDurationMonths} is a {@code Long} and not an {@code Integer}
+ * because {@code employees.contract_duration_months} is {@code int(10)
+ * unsigned}: its range does not fit a signed {@code int}, so MariaDB
+ * Connector/J hands back a {@code Long}. The same reason
+ * {@link com.workin.legacy.LegacyJdbcValues} reads every numeric column as one.
  */
 public record Employee(
 		long id, long companyId, String companyName, String employeeCode, String employeeName,
 		String phone, String countryCode, boolean active, String hireDate, String createdAt,
-		String photoUrl, Integer contractDurationMonths, String branchName, String departmentName,
+		String photoUrl, Long contractDurationMonths, String branchName, String departmentName,
 		String jobTitleName, String shiftName, BigDecimal basicSalary) {
 
 	public String branchLabel() {
